Organize files by components
